refactor: ListCollectionViewCell과 GridCollectionViewCell 중복 코드 파일로 분리
