+++ OpenMarket/OpenMarket/CollectionViewCell/GridCollectionViewCell.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class GridCollectionViewCell: UICollectionViewCell {
+class GridCollectionViewCell: ItemCollectionViewCell {
     
     // MARK: init
     
@@ -22,38 +22,6 @@ class GridCollectionViewCell: UICollectionViewCell {
     }
     
     // MARK: Properties
-    
-    let productThumnail: UIImageView = {
-        let image = UIImageView()
-        image.contentMode = .scaleAspectFit
-        image.translatesAutoresizingMaskIntoConstraints = false
-        return image
-    }()
-    
-    let productName: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    let productPrice: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    let bargainPrice: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    let productStockQuntity: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-
 
     private let totalGridStackView: UIStackView = {
         let stackView = UIStackView()
